+++ rdf/parse.go
@@ -1,113 +1,12 @@
-/*
- * Copyright 2015 DGraph Labs, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * 		http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package rdf
 
 import (
-	"bytes"
 	"fmt"
 	"strconv"
-	"strings"
-	"time"
 	"unicode"
 
 	p "github.com/dgraph-io/dgraph/parser"
-	"github.com/dgraph-io/dgraph/uid"
-	"github.com/dgraph-io/dgraph/x"
 )
-
-// Gets the uid corresponding to an xid from the posting list which stores the
-// mapping.
-func getUid(xid string) (uint64, error) {
-	// If string represents a UID, convert to uint64 and return.
-	if strings.HasPrefix(xid, "_uid_:") {
-		return strconv.ParseUint(xid[6:], 0, 64)
-	}
-	// Get uid from posting list in UidStore.
-	return uid.Get(xid)
-}
-
-// ToEdge is useful when you want to find the UID corresponding to XID for
-// just one edge. The method doesn't automatically generate a UID for an XID.
-func (nq NQuad) ToEdge() (result x.DirectedEdge, rerr error) {
-	sid, err := getUid(nq.Subject)
-	if err != nil {
-		return result, err
-	}
-
-	result.Entity = sid
-	// An edge can have an id or a value.
-	if len(nq.ObjectId) > 0 {
-		oid, err := getUid(nq.ObjectId)
-		if err != nil {
-			return result, err
-		}
-		result.ValueId = oid
-	} else {
-		result.Value = nq.ObjectValue
-	}
-	result.Attribute = nq.Predicate
-	result.Source = nq.Label
-	result.Timestamp = time.Now()
-	return result, nil
-}
-
-func toUid(xid string, xidToUID map[string]uint64) (uid uint64, rerr error) {
-	if id, present := xidToUID[xid]; present {
-		return id, nil
-	}
-
-	if !strings.HasPrefix(xid, "_uid_:") {
-		return 0, fmt.Errorf("Unable to find xid: %v", xid)
-	}
-	return strconv.ParseUint(xid[6:], 0, 64)
-}
-
-// ToEdgeUsing determines the UIDs for the provided XIDs and populates the
-// xidToUid map.
-func (nq NQuad) ToEdgeUsing(
-	xidToUID map[string]uint64) (result x.DirectedEdge, rerr error) {
-	uid, err := toUid(nq.Subject, xidToUID)
-	if err != nil {
-		return result, err
-	}
-	result.Entity = uid
-
-	if len(nq.ObjectId) == 0 {
-		result.Value = nq.ObjectValue
-	} else {
-		uid, err = toUid(nq.ObjectId, xidToUID)
-		if err != nil {
-			return result, err
-		}
-		result.ValueId = uid
-	}
-	result.Attribute = nq.Predicate
-	result.Source = nq.Label
-	result.Timestamp = time.Now()
-	return result, nil
-}
-
-// This function is used to extract an IRI from an IRIREF.
-func stripBracketsIfPresent(val string) string {
-	if val[0] != '<' && val[len(val)-1] != '>' {
-		return val
-	}
-	return val[1 : len(val)-1]
-}
 
 var pSubject = p.ParseFunc(func(c p.Context) p.Context {
 	var i int
@@ -338,75 +237,3 @@ var pNQuadStatement = p.ParseFunc(func(c p.Context) p.Context {
 	c = c.Parse(pByte('.'))
 	return c.WithValue(ret)
 })
-
-func Parse(line string) (rnq NQuad, err error) {
-	s := p.NewByteStream(bytes.NewBufferString(line))
-	c := p.NewContext(s).Parse(pNQuadStatement)
-	rnq = c.Value().(NQuad)
-	err = c.Err()
-	return
-}
-
-// // Parse parses a mutation string and returns the NQuad representation for it.
-// func Parse(line string) (rnq NQuad, rerr error) {
-// 	l := &lex.Lexer{}
-// 	l.Init(line)
-
-// 	go run(l)
-// 	var oval string
-// 	var vend bool
-// 	// We read items from the l.Items channel to which the lexer sends items.
-// 	for item := range l.Items {
-// 		if item.Typ == itemSubject {
-// 			rnq.Subject = stripBracketsIfPresent(item.Val)
-// 		}
-// 		if item.Typ == itemPredicate {
-// 			rnq.Predicate = stripBracketsIfPresent(item.Val)
-// 		}
-// 		if item.Typ == itemObject {
-// 			rnq.ObjectId = stripBracketsIfPresent(item.Val)
-// 		}
-// 		if item.Typ == itemLiteral {
-// 			oval = item.Val
-// 		}
-// 		if item.Typ == itemLanguage {
-// 			rnq.Predicate += "." + item.Val
-// 		}
-// 		if item.Typ == itemObjectType {
-// 			// TODO: Strictly parse common types like integers, floats etc.
-// 			if len(oval) == 0 {
-// 				log.Fatalf(
-// 					"itemObject should be emitted before itemObjectType. Input: [%s]",
-// 					line)
-// 			}
-// 			oval += "@@" + stripBracketsIfPresent(item.Val)
-// 		}
-// 		if item.Typ == lex.ItemError {
-// 			return rnq, fmt.Errorf(item.Val)
-// 		}
-// 		if item.Typ == itemValidEnd {
-// 			vend = true
-// 		}
-// 		if item.Typ == itemLabel {
-// 			rnq.Label = stripBracketsIfPresent(item.Val)
-// 		}
-// 	}
-// 	if !vend {
-// 		return rnq, fmt.Errorf("Invalid end of input. Input: [%s]", line)
-// 	}
-// 	if len(oval) > 0 {
-// 		rnq.ObjectValue = []byte(oval)
-// 	}
-// 	if len(rnq.Subject) == 0 || len(rnq.Predicate) == 0 {
-// 		return rnq, fmt.Errorf("Empty required fields in NQuad. Input: [%s]", line)
-// 	}
-// 	if len(rnq.ObjectId) == 0 && rnq.ObjectValue == nil {
-// 		return rnq, fmt.Errorf("No Object in NQuad. Input: [%s]", line)
-// 	}
-
-// 	return rnq, nil
-// }
-
-func isNewline(r rune) bool {
-	return r == '\n' || r == '\r'
-}
